Make the makefile ccache-friendly
In order to use ccache with the provided makefile, one would use
CXX="ccache g++" for example. The problem is, the generated object
directories will contain a space, and make will become confused because
it doesn't like spaces. Now, only the second part of the compiler name
is used for the directory name (example: when using "ccache g++", the
directory will be name "obj/g++-<type>", not "obj/ccache g++-<type>").